+++ Source/ESBlockingPlugin.m
@@ -630,8 +630,8 @@
 									AILocalizedString(@"Are you sure you want to block %@?",nil) :
 									AILocalizedString(@"Are you sure you want to unblock %@?",nil));
 				
-				NSString *questionQualifier = [NSString stringWithFormat:AILocalizedString(@"%d contacts", nil), 
-											   activeChatInWindow.containedObjects.count];
+				NSString *questionQualifier = [NSString stringWithFormat:AILocalizedString(@"%lu contacts", nil), 
+											   (unsigned long)activeChatInWindow.containedObjects.count];
 				
 				if(activeChatInWindow.containedObjects.count == 1) {
 					questionQualifier = [[activeChatInWindow.containedObjects objectAtIndex:0] displayName];
